+++ stroom-query-api/src/main/java/stroom/query/api/v2/ExpressionOperator.java
@@ -284,40 +284,21 @@ public final class ExpressionOperator extends ExpressionItem {
         }
 
         /**
-         * A convenience function for adding dictionary terms in one go, the parameters should read fairly clearly
-         *
-         * @param field      The field name
-         * @param condition  The condition to apply to the value
-         * @param dictionary The dictionary
-         * @return this builder, with the completed term added.
-         */
-        public Builder addDictionaryTerm(final String field,
-                                         final ExpressionTerm.Condition condition,
-                                         final DocRef dictionary) {
-            return addTerm(new ExpressionTerm.Builder()
-                    .field(field)
-                    .condition(condition)
-                    .dictionary(dictionary)
-                    .build());
-        }
-
-        /**
          * A convenience function for adding docRef terms in one go, the parameters should read fairly clearly
          *
-         * @param field      The field name
-         * @param condition  The condition to apply to the value
-         * @param docRef The docRef
+         * @param field     The field name
+         * @param condition The condition to apply to the value
+         * @param docRef    The docRef
          * @return this builder, with the completed term added.
          */
         public Builder addDocRefTerm(final String field,
-                                         final ExpressionTerm.Condition condition,
-                                         final DocRef docRef) {
+                                     final ExpressionTerm.Condition condition,
+                                     final DocRef docRef) {
             return addTerm(new ExpressionTerm.Builder()
                     .field(field)
                     .condition(condition)
                     .docRef(docRef)
                     .build());
         }
-
     }
 }