+++ stroom-query-common/src/main/java/stroom/query/common/v2/FlatResultCreator.java
@@ -367,8 +367,6 @@ public class FlatResultCreator implements ResultCreator {
                     }
                 }
                 final TablePayload payload = (TablePayload) tableCoprocessor.createPayload();
-
-                tablePayloadHandler.clear();
                 tablePayloadHandler.addQueue(payload.getQueue());
             }