+++ stroom-query-common/src/main/java/stroom/query/common/v2/SearchResponseCreatorCache.java
@@ -17,7 +17,6 @@ public interface SearchResponseCreatorCache {
 
     /**
      * Remove an entry from the cache for {@link Key} key
-     * @param key The key to remove
      */
     void remove(final Key key);